+++ CHANGELOG.md
@@ -32,6 +32,8 @@
 - New #44: Add `pinChatMessage`, `unpinChatMessage` and `unpinAllChatMessages` methods.
 - New #46: Add `leaveChat` method.
 - New #47: Add `getChatAdministrators`, `getChatMemberCount` and `getChatMember` methods.
+- New #49: Add `createNewStickerSet` method, `InputSticker` type, `StickerType` and `StickerFormat` classes with
+  constants, and `MaskPosition::toRequestArray()` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.